requirements.txt
fastapi==0.135.3
groq==1.1.2
Pillow==12.2.0
pydantic==2.12.5
pyngrok==7.5.1
python-dotenv==1.2.2
python_bcrypt==0.3.2
rembg==2.0.74
slowapi==0.1.9
starlette==1.0.0
uvicorn==0.44.0
services==0.1.1